# Conflicts:
#	src/main/java/com/mercadolibre/w4g9projetofinal/dtos/request/AdvertiseRequestDTO.java
#	src/main/java/com/mercadolibre/w4g9projetofinal/service/InboundOrderService.java
#	src/main/java/com/mercadolibre/w4g9projetofinal/util/Instantiation.java

+++ src/main/java/com/mercadolibre/w4g9projetofinal/dtos/response/BatchResponseDTO.java
@@ -17,7 +17,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BatchResponseDTO {
-    private int batchNumber;
+    private Long batchNumber;
     private Long advertiseId;
     private float currentTemperature;
     private float minimumTemperature;
